sentry: invalid MAC error - log as TRACE
The error is happening if you delete <datadir>/nodekey and the old peers try to reconnect.
They use a public key of an old nodekey to encrypt their AuthMessage, and it will fail the MAC check with the new nodekey.
The behaviour is the same in erigon, and erigon uses logger.Trace.