cmd: fix human readable size node options (#2048)
CLI11 captures option values by references for writing when `cli.parse` is called. When `add_node_options()` exited, the strings were out of scope and the captured references became dangling. That lead to a crash on GCC 13 Release builds. If there was no crash, the option values were not applied either (it was always using default values).

Refactoring:
* add `add_option_human_size` helper to work with `size_t&` performing "human readable" conversions as needed.

Change limits:
* `--chaindata.growthsize` from 64MB to 32MB, because it must be smaller than the `--chaindata.maxsize` which was at 32MB min. It doesn't make sense to have `--chaindata.growthsize` > `--chaindata.maxsize`;
* change `--chaindata.pagesize` max from 65KB to 64KB, because it has to be a power of 2.