cmd: support version update in schema command of snapshots tool (#2292)
This PR adds force_update option to schema subcommand of snapshots command-line utility to update the database schema to the current Silkworm code version. This is useful when upgrading the code between compatible versions or if at some point embedded versioning was wrong (i.e. not correctly updated to match Erigon's one).